+++ fireplace/cards/icecrown/__init__.py
@@ -1,1 +1,1 @@
-from .collectible import *
+from .icecrown import *
