+++ fireplace/cards/icecrown/icecrown.py
@@ -1,0 +1,2308 @@
+
+"""
+Card Definitions for Knights of the Frozen Throne
+Part of Fireplace: The Hearthstone simulator in Python
+"""
+
+from ..utils import *
+
+###############################################################################
+##                                                                           ##
+##                                  Neutral                                  ##
+##                                                                           ##
+###############################################################################
+
+
+# class ICCA01_007:
+# 	"""
+# 	Eager Rogue - (Minion)
+# 	Deals no damage while attacking.
+# 	https://hearthstone.gamepedia.com/Eager_Rogue
+# 	"""
+# 	pass
+
+
+# class ICCA01_009:
+# 	"""
+# 	Needy Hunter - (Minion)
+# 	Whenever this minion attacks, add a Coin to your hand.
+# 	https://hearthstone.gamepedia.com/Needy_Hunter
+# 	"""
+# 	pass
+
+
+# class ICCA01_011:
+# 	"""
+# 	Warlock on Fire - (Minion)
+# 	At the end of your turn, deal 1 damage to this minion.
+# 	https://hearthstone.gamepedia.com/Warlock_on_Fire
+# 	"""
+# 	pass
+
+
+# class ICCA04_008p:
+# 	"""
+# 	Frost Breath - (HeroPower)
+# 	Hero Power Permanently Freeze all enemy minions.
+# 	https://hearthstone.gamepedia.com/Frost_Breath
+# 	"""
+# 	pass
+
+
+# class ICCA04_009p:
+# 	"""
+# 	Frost Breath - (HeroPower)
+# 	Triggered Power At 20 Health, transforms all enemy minions into blocks of ice.
+# 	https://hearthstone.gamepedia.com/Frost_Breath
+# 	"""
+# 	pass
+
+
+# class ICCA04_010p:
+# 	"""
+# 	Frost Breath - (HeroPower)
+# 	Triggered Power At 10 Health, transforms all enemy minions into blocks of ice.
+# 	https://hearthstone.gamepedia.com/Frost_Breath
+# 	"""
+# 	pass
+
+
+# class ICCA04_011p:
+# 	"""
+# 	Ice Claw - (HeroPower)
+# 	Hero Power Deal $2 damage.
+# 	https://hearthstone.gamepedia.com/Ice_Claw
+# 	"""
+# 	pass
+
+
+# class ICCA05_002e:
+# 	"""
+# 	Bitten! - (Enchantment)
+# 	+2/+2. Can't be bitten again.
+# 	https://hearthstone.gamepedia.com/Bitten%21
+# 	"""
+# 	pass
+
+
+# class ICCA05_002p:
+# 	"""
+# 	Vampiric Bite - (HeroPower)
+# 	Hero Power Give a non-Vampire +2/+2. You  use this.
+# 	https://hearthstone.gamepedia.com/Vampiric_Bite
+# 	"""
+# 	pass
+
+
+# class ICCA05_004p:
+# 	"""
+# 	Vampiric Leech - (HeroPower)
+# 	Hero Power Lifesteal Deal $3 damage.
+# 	https://hearthstone.gamepedia.com/Vampiric_Leech
+# 	"""
+# 	pass
+
+
+# class ICCA06_002p:
+# 	"""
+# 	Skeletal Reconstruction - (HeroPower)
+# 	Hero Power Restore your hero to full Health.
+# 	https://hearthstone.gamepedia.com/Skeletal_Reconstruction
+# 	"""
+# 	pass
+
+
+# class ICCA06_004:
+# 	"""
+# 	Bone Storm - (Spell)
+# 	Deal $1 damage to all enemies. Draw a card.
+# 	https://hearthstone.gamepedia.com/Bone_Storm
+# 	"""
+# 	pass
+
+
+# class ICCA06_005:
+# 	"""
+# 	Bone Spike - (Minion)
+# 	At the start of your turn, deal 15 damage to the enemy hero.
+# 	https://hearthstone.gamepedia.com/Bone_Spike
+# 	"""
+# 	pass
+
+
+# class ICCA07_002p:
+# 	"""
+# 	Mad Science - (HeroPower)
+# 	Passive Hero Power All Secrets cost (0).
+# 	https://hearthstone.gamepedia.com/Mad_Science
+# 	"""
+# 	pass
+
+
+# class ICCA07_003p:
+# 	"""
+# 	Madder Science - (HeroPower)
+# 	Passive Hero Power All weapons cost (1).
+# 	https://hearthstone.gamepedia.com/Madder_Science
+# 	"""
+# 	pass
+
+
+# class ICCA07_004:
+# 	"""
+# 	Growing Ooze - (Minion)
+# 	At the start of your turn, gain +1/+1.
+# 	https://hearthstone.gamepedia.com/Growing_Ooze
+# 	"""
+# 	pass
+
+
+# class ICCA07_004e:
+# 	"""
+# 	Grow - (Enchantment)
+# 	The ooze is growing.
+# 	https://hearthstone.gamepedia.com/Grow
+# 	"""
+# 	pass
+
+
+# class ICCA07_005p:
+# 	"""
+# 	Maddest Science - (HeroPower)
+# 	Passive Hero Power All cards cost (5). Because SCIENCE!
+# 	https://hearthstone.gamepedia.com/Maddest_Science
+# 	"""
+# 	pass
+
+
+# class ICCA07_008:
+# 	"""
+# 	Festergut - (Minion)
+# 	At the end of your turn, deal 3 damage to all enemy minions.
+# 	https://hearthstone.gamepedia.com/Festergut
+# 	"""
+# 	pass
+
+
+# class ICCA08_002p:
+# 	"""
+# 	The Scourge - (HeroPower)
+# 	Hero Power Summon a 2/2 Ghoul.
+# 	https://hearthstone.gamepedia.com/The_Scourge
+# 	"""
+# 	pass
+
+
+# class ICCA08_017:
+# 	"""
+# 	Val'kyr Shadowguard - (Minion)
+# 	At the start of your turn, destroy a random enemy minion and this one.
+# 	https://hearthstone.gamepedia.com/Val%27kyr_Shadowguard
+# 	"""
+# 	pass
+
+
+# class ICCA08_022e2:
+# 	"""
+# 	Join the Ranks - (Enchantment)
+# 	When this dies, resummon it for your oponent.
+# 	https://hearthstone.gamepedia.com/Join_the_Ranks
+# 	"""
+# 	pass
+
+
+# class ICCA08_032p:
+# 	"""
+# 	Harvest of Souls - (HeroPower)
+# 	Passive Hero Power You are Immune if you control a Trapped Soul.
+# 	https://hearthstone.gamepedia.com/Harvest_of_Souls
+# 	"""
+# 	pass
+
+
+# class ICCA09_001p:
+# 	"""
+# 	Blood Rune - (HeroPower)
+# 	Passive Hero Power Can only take damage from weapons.
+# 	https://hearthstone.gamepedia.com/Blood_Rune
+# 	"""
+# 	pass
+
+
+# class ICCA09_001t1:
+# 	"""
+# 	Blood Beast - (Minion)
+# 	At the start of your turn, restore 3 Health to your hero.
+# 	https://hearthstone.gamepedia.com/Blood_Beast
+# 	"""
+# 	pass
+
+
+# class ICCA10_001:
+# 	"""
+# 	Valithria Dreamwalker - (Minion)
+# 	Can't attack or be attacked while damaged.
+# 	https://hearthstone.gamepedia.com/Valithria_Dreamwalker
+# 	"""
+# 	pass
+
+
+# class ICCA10_009p:
+# 	"""
+# 	Whisper of Death - (HeroPower)
+# 	Hero Power Damage all enemy minions until they have 1 Health.
+# 	https://hearthstone.gamepedia.com/Whisper_of_Death
+# 	"""
+# 	pass
+
+
+# class ICCA11_001:
+# 	"""
+# 	Skeletal Knight - (Minion)
+# 	Deathrattle: Add a  Knights of the Frozen Throne card to your opponent's hand.
+# 	https://hearthstone.gamepedia.com/Skeletal_Knight
+# 	"""
+# 	deathrattle = None
+# 	pass
+
+
+# class ICC_018:
+# 	"""
+# 	Phantom Freebooter - (Minion)
+# 	Battlecry: Gain stats equal to your weapon's.
+# 	https://hearthstone.gamepedia.com/Phantom_Freebooter
+# 	"""
+# 	play = None
+# 	pass
+
+
+# class ICC_019:
+# 	"""
+# 	Skelemancer - (Minion)
+# 	Deathrattle: If it's your opponent's turn, summon an 8/8 Skeleton.
+# 	https://hearthstone.gamepedia.com/Skelemancer
+# 	"""
+# 	deathrattle = None
+# 	pass
+
+
+class ICC_025:
+	"""
+	Rattling Rascal - (Minion)
+	Battlecry: Summon a 5/5 Skeleton. Deathrattle: Summon one for your opponent.
+	https://hearthstone.gamepedia.com/Rattling_Rascal
+	"""
+	play = Summon(CONTROLLER, "ICC_025t")
+	deathrattle = Summon(OPPONENT, "ICC_025t")
+
+
+
+# class ICC_026:
+# 	"""
+# 	Grim Necromancer - (Minion)
+# 	Battlecry: Summon two 1/1 Skeletons.
+# 	https://hearthstone.gamepedia.com/Grim_Necromancer
+# 	"""
+# 	play = None
+# 	pass
+
+
+# class ICC_027:
+# 	"""
+# 	Bone Drake - (Minion)
+# 	Deathrattle: Add a random Dragon to your hand.
+# 	https://hearthstone.gamepedia.com/Bone_Drake
+# 	"""
+# 	deathrattle = None
+# 	pass
+
+
+# class ICC_028:
+# 	"""
+# 	Sunborne Val'kyr - (Minion)
+# 	Battlecry: Give adjacent minions +2 Health.
+# 	https://hearthstone.gamepedia.com/Sunborne_Val%27kyr
+# 	"""
+# 	play = None
+# 	pass
+
+
+# class ICC_028e:
+# 	"""
+# 	Blessing of the Val'kyr - (Enchantment)
+# 	+2 Health.
+# 	https://hearthstone.gamepedia.com/Blessing_of_the_Val%27kyr
+# 	"""
+# 	pass
+
+
+# class ICC_029:
+# 	"""
+# 	Cobalt Scalebane - (Minion)
+# 	At the end of your turn, give another random friendly minion +3 Attack.
+# 	https://hearthstone.gamepedia.com/Cobalt_Scalebane
+# 	"""
+# 	pass
+
+
+# class ICC_029e:
+# 	"""
+# 	Dragonscales - (Enchantment)
+# 	Attack increased.
+# 	https://hearthstone.gamepedia.com/Dragonscales
+# 	"""
+# 	pass
+
+
+# class ICC_031:
+# 	"""
+# 	Night Howler - (Minion)
+# 	Whenever this minion takes damage, gain +2 Attack.
+# 	https://hearthstone.gamepedia.com/Night_Howler
+# 	"""
+# 	pass
+
+
+# class ICC_031e:
+# 	"""
+# 	Awooooo! - (Enchantment)
+# 	This minion has increased Attack.
+# 	https://hearthstone.gamepedia.com/Awooooo%21
+# 	"""
+# 	pass
+
+
+# class ICC_039e:
+# 	"""
+# 	Convinced - (Enchantment)
+# 	Stats changed to 3/3.
+# 	https://hearthstone.gamepedia.com/Convinced
+# 	"""
+# 	pass
+
+
+# class ICC_056e:
+# 	"""
+# 	Frozen Blood - (Enchantment)
+# 	+3/+3.
+# 	https://hearthstone.gamepedia.com/Frozen_Blood
+# 	"""
+# 	pass
+
+
+# class ICC_067:
+# 	"""
+# 	Vryghoul - (Minion)
+# 	Deathrattle: If it's your opponent's turn, summon a 2/2 Ghoul.
+# 	https://hearthstone.gamepedia.com/Vryghoul
+# 	"""
+# 	deathrattle = None
+# 	pass
+
+
+# class ICC_092:
+# 	"""
+# 	Acherus Veteran - (Minion)
+# 	Battlecry: Give a friendly minion +1 Attack.
+# 	https://hearthstone.gamepedia.com/Acherus_Veteran
+# 	"""
+# 	play = None
+# 	pass
+
+
+# class ICC_092e:
+# 	"""
+# 	Veteran's Favor - (Enchantment)
+# 	+1 Attack.
+# 	https://hearthstone.gamepedia.com/Veteran%27s_Favor
+# 	"""
+# 	pass
+
+
+# class ICC_093:
+# 	"""
+# 	Tuskarr Fisherman - (Minion)
+# 	Battlecry: Give a friendly minion Spell Damage +1.
+# 	https://hearthstone.gamepedia.com/Tuskarr_Fisherman
+# 	"""
+# 	play = None
+# 	pass
+
+
+# class ICC_094:
+# 	"""
+# 	Fallen Sun Cleric - (Minion)
+# 	Battlecry: Give a friendly minion +1/+1.
+# 	https://hearthstone.gamepedia.com/Fallen_Sun_Cleric
+# 	"""
+# 	play = None
+# 	pass
+
+
+# class ICC_094e:
+# 	"""
+# 	Cleric's Blessing - (Enchantment)
+# 	+1/+1.
+# 	https://hearthstone.gamepedia.com/Cleric%27s_Blessing
+# 	"""
+# 	pass
+
+
+# class ICC_096:
+# 	"""
+# 	Furnacefire Colossus - (Minion)
+# 	Battlecry: Discard all weapons from your hand and gain their stats.
+# 	https://hearthstone.gamepedia.com/Furnacefire_Colossus
+# 	"""
+# 	play = None
+# 	pass
+
+
+# class ICC_096e:
+# 	"""
+# 	Me Bigger - (Enchantment)
+# 	Increased stats.
+# 	https://hearthstone.gamepedia.com/Me_Bigger
+# 	"""
+# 	pass
+
+
+# class ICC_097:
+# 	"""
+# 	Grave Shambler - (Minion)
+# 	Whenever your weapon is destroyed, gain +1/+1.
+# 	https://hearthstone.gamepedia.com/Grave_Shambler
+# 	"""
+# 	pass
+
+
+# class ICC_097e:
+# 	"""
+# 	Armed and Dangerous - (Enchantment)
+# 	Increased stats.
+# 	https://hearthstone.gamepedia.com/Armed_and_Dangerous
+# 	"""
+# 	pass
+
+
+# class ICC_098:
+# 	"""
+# 	Tomb Lurker - (Minion)
+# 	Battlecry: Add a random Deathrattle minion that died this game to your hand.
+# 	https://hearthstone.gamepedia.com/Tomb_Lurker
+# 	"""
+# 	play = None
+# 	pass
+
+
+# class ICC_099:
+# 	"""
+# 	Ticking Abomination - (Minion)
+# 	Deathrattle: Deal 5 damage to your minions.
+# 	https://hearthstone.gamepedia.com/Ticking_Abomination
+# 	"""
+# 	deathrattle = None
+# 	pass
+
+
+# class ICC_235e:
+# 	"""
+# 	Shadow Essence - (Enchantment)
+# 	Attack and Health set to 5.
+# 	https://hearthstone.gamepedia.com/Shadow_Essence
+# 	"""
+# 	pass
+
+
+# class ICC_257:
+# 	"""
+# 	Corpse Raiser - (Minion)
+# 	Battlecry: Give a friendly minion "Deathrattle:   Resummon this minion."
+# 	https://hearthstone.gamepedia.com/Corpse_Raiser
+# 	"""
+# 	play = None
+# 	pass
+
+
+# class ICC_257e:
+# 	"""
+# 	Ready to Return - (Enchantment)
+# 	Deathrattle: Resummon this minion.
+# 	https://hearthstone.gamepedia.com/Ready_to_Return
+# 	"""
+# 	pass
+
+
+# class ICC_314:
+# 	"""
+# 	The Lich King - (Minion)
+# 	Taunt At the end of your turn, add a random Death Knight card to your hand.
+# 	https://hearthstone.gamepedia.com/The_Lich_King
+# 	"""
+# 	pass
+
+
+# class ICC_314t1e:
+# 	"""
+# 	Trapped Soul - (Enchantment)
+# 	{0}
+# 	https://hearthstone.gamepedia.com/Trapped_Soul
+# 	"""
+# 	pass
+
+
+# class ICC_314t7e:
+# 	"""
+# 	Anti-Magic Shell - (Enchantment)
+# 	+2/+2 and "Can't be targeted by spells or Hero Powers."
+# 	https://hearthstone.gamepedia.com/Anti-Magic_Shell
+# 	"""
+# 	pass
+
+
+# class ICC_450e:
+# 	"""
+# 	Bloodthirsty - (Enchantment)
+# 	Increased stats.
+# 	https://hearthstone.gamepedia.com/Bloodthirsty
+# 	"""
+# 	pass
+
+
+# class ICC_466:
+# 	"""
+# 	Saronite Chain Gang - (Minion)
+# 	Taunt Battlecry: Summon a copy of this minion.
+# 	https://hearthstone.gamepedia.com/Saronite_Chain_Gang
+# 	"""
+# 	play = None
+# 	pass
+
+
+# class ICC_467:
+# 	"""
+# 	Deathspeaker - (Minion)
+# 	Battlecry: Give a friendly minion Immune this turn.
+# 	https://hearthstone.gamepedia.com/Deathspeaker
+# 	"""
+# 	play = None
+# 	pass
+
+
+# class ICC_467e:
+# 	"""
+# 	Deathward - (Enchantment)
+# 	Immune this turn.
+# 	https://hearthstone.gamepedia.com/Deathward
+# 	"""
+# 	pass
+
+
+# class ICC_468:
+# 	"""
+# 	Wretched Tiller - (Minion)
+# 	Whenever this minion attacks, deal 2 damage to_the enemy hero.
+# 	https://hearthstone.gamepedia.com/Wretched_Tiller
+# 	"""
+# 	pass
+
+
+# class ICC_481p:
+# 	"""
+# 	Transmute Spirit - (HeroPower)
+# 	Hero Power: Transform a friendly minion into a random one that costs (1) more.
+# 	https://hearthstone.gamepedia.com/Transmute_Spirit
+# 	"""
+# 	pass
+
+
+# class ICC_700:
+# 	"""
+# 	Happy Ghoul - (Minion)
+# 	Costs (0) if your hero was healed this turn.
+# 	https://hearthstone.gamepedia.com/Happy_Ghoul
+# 	"""
+# 	pass
+
+
+# class ICC_701:
+# 	"""
+# 	Skulking Geist - (Minion)
+# 	Battlecry: Destroy all 1-Cost spells in both hands and decks.
+# 	https://hearthstone.gamepedia.com/Skulking_Geist
+# 	"""
+# 	play = None
+# 	pass
+
+
+# class ICC_702:
+# 	"""
+# 	Shallow Gravedigger - (Minion)
+# 	Deathrattle: Add a random Deathrattle minion to your_hand.
+# 	https://hearthstone.gamepedia.com/Shallow_Gravedigger
+# 	"""
+# 	deathrattle = None
+# 	pass
+
+
+# class ICC_705:
+# 	"""
+# 	Bonemare - (Minion)
+# 	Battlecry: Give a friendly minion +4/+4 and Taunt.
+# 	https://hearthstone.gamepedia.com/Bonemare
+# 	"""
+# 	play = None
+# 	pass
+
+
+# class ICC_705e:
+# 	"""
+# 	Bonemare's Boon - (Enchantment)
+# 	+4/+4 and Taunt.
+# 	https://hearthstone.gamepedia.com/Bonemare%27s_Boon
+# 	"""
+# 	pass
+
+
+# class ICC_706:
+# 	"""
+# 	Nerubian Unraveler - (Minion)
+# 	Spells cost (2) more.
+# 	https://hearthstone.gamepedia.com/Nerubian_Unraveler
+# 	"""
+# 	pass
+
+
+# class ICC_807e:
+# 	"""
+# 	Strongshell - (Enchantment)
+# 	+2/+2.
+# 	https://hearthstone.gamepedia.com/Strongshell
+# 	"""
+# 	pass
+
+
+# class ICC_810:
+# 	"""
+# 	Deathaxe Punisher - (Minion)
+# 	Battlecry: Give a random Lifesteal minion in your hand +2/+2.
+# 	https://hearthstone.gamepedia.com/Deathaxe_Punisher
+# 	"""
+# 	play = None
+# 	pass
+
+
+# class ICC_810e:
+# 	"""
+# 	Bloodthirsty - (Enchantment)
+# 	+2/+2 from Deathaxe Punisher.
+# 	https://hearthstone.gamepedia.com/Bloodthirsty
+# 	"""
+# 	pass
+
+
+# class ICC_812:
+# 	"""
+# 	Meat Wagon - (Minion)
+# 	Deathrattle: Summon a minion from your deck with less Attack than this minion.
+# 	https://hearthstone.gamepedia.com/Meat_Wagon
+# 	"""
+# 	deathrattle = None
+# 	pass
+
+
+# class ICC_828p:
+# 	"""
+# 	Build-A-Beast - (HeroPower)
+# 	Hero Power Craft a custom Zombeast.
+# 	https://hearthstone.gamepedia.com/Build-A-Beast
+# 	"""
+# 	pass
+
+
+# class ICC_829p:
+# 	"""
+# 	The Four Horsemen - (HeroPower)
+# 	Hero Power Summon a 2/2 Horseman. If you have all 4, destroy the enemy hero.
+# 	https://hearthstone.gamepedia.com/The_Four_Horsemen
+# 	"""
+# 	pass
+
+
+# class ICC_831p:
+# 	"""
+# 	Siphon Life - (HeroPower)
+# 	Hero Power Lifesteal Deal $3 damage.
+# 	https://hearthstone.gamepedia.com/Siphon_Life
+# 	"""
+# 	pass
+
+
+# class ICC_832e:
+# 	"""
+# 	Fangs - (Enchantment)
+# 	+3 Attack this turn.
+# 	https://hearthstone.gamepedia.com/Fangs
+# 	"""
+# 	pass
+
+
+# class ICC_832p:
+# 	"""
+# 	Plague Lord - (HeroPower)
+# 	Hero Power Choose One - +3 Attack this turn; or Gain 3 Armor.
+# 	https://hearthstone.gamepedia.com/Plague_Lord
+# 	"""
+# 	pass
+
+
+# class ICC_833e:
+# 	"""
+# 	Frost Lich - (Enchantment)
+# 	Your Elementals have Lifesteal.
+# 	https://hearthstone.gamepedia.com/Frost_Lich
+# 	"""
+# 	pass
+
+
+# class ICC_833e2:
+# 	"""
+# 	Icy Veins - (Enchantment)
+# 	Frost Lich Jaina is granting your Elementals Lifesteal.
+# 	https://hearthstone.gamepedia.com/Icy_Veins
+# 	"""
+# 	pass
+
+
+# class ICC_833h:
+# 	"""
+# 	Icy Touch - (HeroPower)
+# 	Hero Power  Deal $1 damage. If this kills a minion, summon a Water Elemental.
+# 	https://hearthstone.gamepedia.com/Icy_Touch
+# 	"""
+# 	pass
+
+
+# class ICC_837e:
+# 	"""
+# 	Challenged - (Enchantment)
+# 	Minion cost reduced by (2).
+# 	https://hearthstone.gamepedia.com/Challenged
+# 	"""
+# 	pass
+
+
+# class ICC_841e:
+# 	"""
+# 	Vampiric Bite - (Enchantment)
+# 	Increased Attack.
+# 	https://hearthstone.gamepedia.com/Vampiric_Bite
+# 	"""
+# 	pass
+
+
+# class ICC_849e:
+# 	"""
+# 	Embraced - (Enchantment)
+# 	At the start of the next turn, lose control of this minion.
+# 	https://hearthstone.gamepedia.com/Embraced
+# 	"""
+# 	pass
+
+
+# class ICC_851:
+# 	"""
+# 	Prince Keleseth - (Minion)
+# 	Battlecry: If your deck has_no 2-Cost cards, give_all minions in your deck +1/+1.
+# 	https://hearthstone.gamepedia.com/Prince_Keleseth
+# 	"""
+# 	play = None
+# 	pass
+
+
+# class ICC_851e:
+# 	"""
+# 	Keleseth's Blessing - (Enchantment)
+# 	+1/+1.
+# 	https://hearthstone.gamepedia.com/Keleseth%27s_Blessing
+# 	"""
+# 	pass
+
+
+# class ICC_852:
+# 	"""
+# 	Prince Taldaram - (Minion)
+# 	Battlecry: If your deck has_no 3-Cost cards, transform into a 3/3 copy of a minion.
+# 	https://hearthstone.gamepedia.com/Prince_Taldaram
+# 	"""
+# 	play = None
+# 	pass
+
+
+# class ICC_852e:
+# 	"""
+# 	Taldaram's Visage - (Enchantment)
+# 	3/3.
+# 	https://hearthstone.gamepedia.com/Taldaram%27s_Visage
+# 	"""
+# 	pass
+
+
+# class ICC_853:
+# 	"""
+# 	Prince Valanar - (Minion)
+# 	Battlecry: If your deck has no 4-Cost cards, gain Lifesteal and Taunt.
+# 	https://hearthstone.gamepedia.com/Prince_Valanar
+# 	"""
+# 	play = None
+# 	pass
+
+
+# class ICC_854:
+# 	"""
+# 	Arfus - (Minion)
+# 	Deathrattle: Add a random Death Knight card to your_hand.
+# 	https://hearthstone.gamepedia.com/Arfus
+# 	"""
+# 	deathrattle = None
+# 	pass
+
+
+# class ICC_855:
+# 	"""
+# 	Hyldnir Frostrider - (Minion)
+# 	Battlecry: Freeze your other minions.
+# 	https://hearthstone.gamepedia.com/Hyldnir_Frostrider
+# 	"""
+# 	play = None
+# 	pass
+
+
+# class ICC_900:
+# 	"""
+# 	Necrotic Geist - (Minion)
+# 	Whenever one of your other minions dies, summon a 2/2 Ghoul.
+# 	https://hearthstone.gamepedia.com/Necrotic_Geist
+# 	"""
+# 	pass
+
+
+# class ICC_901:
+# 	"""
+# 	Drakkari Enchanter - (Minion)
+# 	Your end of turn effects trigger twice.
+# 	https://hearthstone.gamepedia.com/Drakkari_Enchanter
+# 	"""
+# 	pass
+
+
+# class ICC_902:
+# 	"""
+# 	Mindbreaker - (Minion)
+# 	Hero Powers are disabled.
+# 	https://hearthstone.gamepedia.com/Mindbreaker
+# 	"""
+# 	pass
+
+
+# class ICC_903t:
+# 	"""
+# 	Bloodthirst - (Enchantment)
+# 	Increased stats.
+# 	https://hearthstone.gamepedia.com/Bloodthirst
+# 	"""
+# 	pass
+
+
+# class ICC_904:
+# 	"""
+# 	Wicked Skeleton - (Minion)
+# 	Battlecry: Gain +1/+1 for_each minion that died_this turn.
+# 	https://hearthstone.gamepedia.com/Wicked_Skeleton
+# 	"""
+# 	play = None
+# 	pass
+
+
+# class ICC_904e:
+# 	"""
+# 	Extra Calcium - (Enchantment)
+# 	Increased stats.
+# 	https://hearthstone.gamepedia.com/Extra_Calcium
+# 	"""
+# 	pass
+
+
+class ICC_911:
+	"""
+	Keening Banshee - (Minion)
+	Whenever you play a card, remove the top 3 cards of_your deck.
+	https://hearthstone.gamepedia.com/Keening_Banshee
+	"""
+	events = Play(CONTROLLER).on(Mill(CONTROLLER, 3))
+
+# class ICC_912:
+# 	"""
+# 	Corpsetaker - (Minion)
+# 	Battlecry: Gain Taunt if your deck has a Taunt minion. Repeat for Divine Shield, Lifesteal, Windfury.
+# 	https://hearthstone.gamepedia.com/Corpsetaker
+# 	"""
+# 	play = None
+# 	pass
+
+###############################################################################
+##                                                                           ##
+##                                  Warrior                                  ##
+##                                                                           ##
+###############################################################################
+
+
+# class ICC_062:
+# 	"""
+# 	Mountainfire Armor - (Minion)
+# 	Deathrattle: If it's your opponent's turn, gain 6 Armor.
+# 	https://hearthstone.gamepedia.com/Mountainfire_Armor
+# 	"""
+# 	deathrattle = None
+# 	pass
+
+
+# class ICC_064:
+# 	"""
+# 	Blood Razor - (Weapon)
+# 	Battlecry and Deathrattle: Deal 1 damage to all_minions.
+# 	https://hearthstone.gamepedia.com/Blood_Razor
+# 	"""
+# 	play = None
+# 	deathrattle = None
+# 	pass
+
+
+# class ICC_091:
+# 	"""
+# 	Dead Man's Hand - (Spell)
+# 	Shuffle a copy of your hand into your deck.
+# 	https://hearthstone.gamepedia.com/Dead_Man%27s_Hand
+# 	"""
+# 	pass
+
+
+# class ICC_238:
+# 	"""
+# 	Animated Berserker - (Minion)
+# 	After you play a minion, deal 1 damage to it.
+# 	https://hearthstone.gamepedia.com/Animated_Berserker
+# 	"""
+# 	pass
+
+
+# class ICC_281:
+# 	"""
+# 	Forge of Souls - (Spell)
+# 	Draw 2 weapons from your deck.
+# 	https://hearthstone.gamepedia.com/Forge_of_Souls
+# 	"""
+# 	pass
+
+
+# class ICC_405:
+# 	"""
+# 	Rotface - (Minion)
+# 	Whenever this minion survives damage, summon a random Legendary minion.
+# 	https://hearthstone.gamepedia.com/Rotface
+# 	"""
+# 	pass
+
+
+# class ICC_408:
+# 	"""
+# 	Val'kyr Soulclaimer - (Minion)
+# 	Whenever this minion survives damage, summon a 2/2 Ghoul.
+# 	https://hearthstone.gamepedia.com/Val%27kyr_Soulclaimer
+# 	"""
+# 	pass
+
+
+# class ICC_450:
+# 	"""
+# 	Death Revenant - (Minion)
+# 	Battlecry: Gain +1/+1 for each damaged minion.
+# 	https://hearthstone.gamepedia.com/Death_Revenant
+# 	"""
+# 	play = None
+# 	pass
+
+
+# class ICC_834:
+# 	"""
+# 	Scourgelord Garrosh - (Hero)
+# 	Battlecry: Equip a 4/3_Shadowmourne that also damages adjacent minions.
+# 	https://hearthstone.gamepedia.com/Scourgelord_Garrosh
+# 	"""
+# 	play = None
+# 	pass
+
+
+# class ICC_834h:
+# 	"""
+# 	Bladestorm - (HeroPower)
+# 	Hero Power  Deal $1 damage to all_minions.
+# 	https://hearthstone.gamepedia.com/Bladestorm
+# 	"""
+# 	pass
+
+
+# class ICC_834w:
+# 	"""
+# 	Shadowmourne - (Weapon)
+# 	Also damages the minions next to whomever your hero_attacks.
+# 	https://hearthstone.gamepedia.com/Shadowmourne
+# 	"""
+# 	pass
+
+
+# class ICC_837:
+# 	"""
+# 	Bring It On! - (Spell)
+# 	Gain 10 Armor. Reduce the Cost of minions in your opponent's hand by (2).
+# 	https://hearthstone.gamepedia.com/Bring_It_On%21
+# 	"""
+# 	pass
+
+###############################################################################
+##                                                                           ##
+##                                  Warlock                                  ##
+##                                                                           ##
+###############################################################################
+
+
+# class ICCA05_020:
+# 	"""
+# 	Bite of the Blood-Queen - (Spell)
+# 	Turn a hero into a Vampire.
+# 	https://hearthstone.gamepedia.com/Bite_of_the_Blood-Queen
+# 	"""
+# 	pass
+
+
+# class ICCA05_021:
+# 	"""
+# 	Blood Essence - (Spell)
+# 	Transform 2 random minions in your opponent's hand into spells.
+# 	https://hearthstone.gamepedia.com/Blood_Essence
+# 	"""
+# 	pass
+
+
+# class ICC_041:
+# 	"""
+# 	Defile - (Spell)
+# 	Deal $1 damage to all minions. If any die, cast this again.
+# 	https://hearthstone.gamepedia.com/Defile
+# 	"""
+# 	pass
+
+
+# class ICC_055:
+# 	"""
+# 	Drain Soul - (Spell)
+# 	Lifesteal Deal $2 damage to a minion.
+# 	https://hearthstone.gamepedia.com/Drain_Soul
+# 	"""
+# 	pass
+
+
+# class ICC_075:
+# 	"""
+# 	Despicable Dreadlord - (Minion)
+# 	At the end of your turn, deal 1 damage to all enemy minions.
+# 	https://hearthstone.gamepedia.com/Despicable_Dreadlord
+# 	"""
+# 	pass
+
+
+# class ICC_206:
+# 	"""
+# 	Treachery - (Spell)
+# 	Choose a friendly minion and give it to_your opponent.
+# 	https://hearthstone.gamepedia.com/Treachery
+# 	"""
+# 	pass
+
+
+# class ICC_218:
+# 	"""
+# 	Howlfiend - (Minion)
+# 	Whenever this minion takes damage, discard a_random card.
+# 	https://hearthstone.gamepedia.com/Howlfiend
+# 	"""
+# 	pass
+
+
+# class ICC_407:
+# 	"""
+# 	Gnomeferatu - (Minion)
+# 	Battlecry: Remove the top card of your opponent's deck.
+# 	https://hearthstone.gamepedia.com/Gnomeferatu
+# 	"""
+# 	play = None
+# 	pass
+
+
+# class ICC_469:
+# 	"""
+# 	Unwilling Sacrifice - (Spell)
+# 	Choose a friendly minion. Destroy it and a random enemy minion.
+# 	https://hearthstone.gamepedia.com/Unwilling_Sacrifice
+# 	"""
+# 	pass
+
+
+# class ICC_831:
+# 	"""
+# 	Bloodreaver Gul'dan - (Hero)
+# 	Battlecry: Summon all friendly Demons that_died_this_game.
+# 	https://hearthstone.gamepedia.com/Bloodreaver_Gul%27dan
+# 	"""
+# 	play = None
+# 	pass
+
+
+# class ICC_841:
+# 	"""
+# 	Blood-Queen Lana'thel - (Minion)
+# 	Lifesteal Has +1 Attack for each card you've discarded this game.
+# 	https://hearthstone.gamepedia.com/Blood-Queen_Lana%27thel
+# 	"""
+# 	pass
+
+
+# class ICC_903:
+# 	"""
+# 	Sanguine Reveler - (Minion)
+# 	Battlecry: Destroy a friendly minion and gain_+2/+2.
+# 	https://hearthstone.gamepedia.com/Sanguine_Reveler
+# 	"""
+# 	play = None
+# 	pass
+
+###############################################################################
+##                                                                           ##
+##                                  Shaman                                   ##
+##                                                                           ##
+###############################################################################
+
+
+# class ICC_056:
+# 	"""
+# 	Cryostasis - (Spell)
+# 	Give a minion +3/+3 and Freeze it.
+# 	https://hearthstone.gamepedia.com/Cryostasis
+# 	"""
+# 	pass
+
+
+# class ICC_058:
+# 	"""
+# 	Brrrloc - (Minion)
+# 	Battlecry: Freeze an_enemy.
+# 	https://hearthstone.gamepedia.com/Brrrloc
+# 	"""
+# 	play = None
+# 	pass
+
+
+# class ICC_078:
+# 	"""
+# 	Avalanche - (Spell)
+# 	Freeze a minion and deal $3 damage to adjacent ones.
+# 	https://hearthstone.gamepedia.com/Avalanche
+# 	"""
+# 	pass
+
+
+# class ICC_088:
+# 	"""
+# 	Voodoo Hexxer - (Minion)
+# 	Taunt Freeze any character damaged by this minion.
+# 	https://hearthstone.gamepedia.com/Voodoo_Hexxer
+# 	"""
+# 	pass
+
+
+# class ICC_089:
+# 	"""
+# 	Ice Fishing - (Spell)
+# 	Draw 2 Murlocs from your deck.
+# 	https://hearthstone.gamepedia.com/Ice_Fishing
+# 	"""
+# 	pass
+
+
+# class ICC_090:
+# 	"""
+# 	Snowfury Giant - (Minion)
+# 	Costs (1) less for each Mana Crystal you've Overloaded this game.
+# 	https://hearthstone.gamepedia.com/Snowfury_Giant
+# 	"""
+# 	pass
+
+
+# class ICC_236:
+# 	"""
+# 	Ice Breaker - (Weapon)
+# 	Destroy any Frozen minion damaged by_this.
+# 	https://hearthstone.gamepedia.com/Ice_Breaker
+# 	"""
+# 	pass
+
+
+# class ICC_289:
+# 	"""
+# 	Moorabi - (Minion)
+# 	Whenever another minion is Frozen, add a copy of it to your hand.
+# 	https://hearthstone.gamepedia.com/Moorabi
+# 	"""
+# 	pass
+
+
+# class ICC_481:
+# 	"""
+# 	Thrall, Deathseer - (Hero)
+# 	Battlecry: Transform your minions into random ones that cost (2) more.
+# 	https://hearthstone.gamepedia.com/Thrall%2C_Deathseer
+# 	"""
+# 	play = None
+# 	pass
+
+###############################################################################
+##                                                                           ##
+##                                   Rogue                                   ##
+##                                                                           ##
+###############################################################################
+
+
+# class ICC_018e:
+# 	"""
+# 	Witty Weaponplay - (Enchantment)
+# 	Increased stats.
+# 	https://hearthstone.gamepedia.com/Witty_Weaponplay
+# 	"""
+# 	pass
+
+
+# class ICC_065:
+# 	"""
+# 	Bone Baron - (Minion)
+# 	Deathrattle: Add two 1/1 Skeletons to your hand.
+# 	https://hearthstone.gamepedia.com/Bone_Baron
+# 	"""
+# 	deathrattle = None
+# 	pass
+
+
+# class ICC_201:
+# 	"""
+# 	Roll the Bones - (Spell)
+# 	Draw a card. If it has Deathrattle, cast this again.
+# 	https://hearthstone.gamepedia.com/Roll_the_Bones
+# 	"""
+# 	pass
+
+
+# class ICC_221:
+# 	"""
+# 	Leeching Poison - (Spell)
+# 	Give your weapon Lifesteal.
+# 	https://hearthstone.gamepedia.com/Leeching_Poison
+# 	"""
+# 	pass
+
+
+# class ICC_233:
+# 	"""
+# 	Doomerang - (Spell)
+# 	Throw your weapon at a minion. It deals its damage, then returns to_your hand.
+# 	https://hearthstone.gamepedia.com/Doomerang
+# 	"""
+# 	pass
+
+
+# class ICC_240:
+# 	"""
+# 	Runeforge Haunter - (Minion)
+# 	During your turn, your weapon doesn't lose Durability.
+# 	https://hearthstone.gamepedia.com/Runeforge_Haunter
+# 	"""
+# 	pass
+
+
+# class ICC_240e:
+# 	"""
+# 	Resilient Weapon - (Enchantment)
+# 	No durability loss.
+# 	https://hearthstone.gamepedia.com/Resilient_Weapon
+# 	"""
+# 	pass
+
+
+# class ICC_809:
+# 	"""
+# 	Plague Scientist - (Minion)
+# 	Combo: Give a friendly minion Poisonous.
+# 	https://hearthstone.gamepedia.com/Plague_Scientist
+# 	"""
+# 	pass
+
+
+# class ICC_811:
+# 	"""
+# 	Lilian Voss - (Minion)
+# 	Battlecry: Replace spells in your hand with random spells
+# 	https://hearthstone.gamepedia.com/Lilian_Voss
+# 	"""
+# 	play = None
+# 	pass
+
+
+# class ICC_827:
+# 	"""
+# 	Valeera the Hollow - (Hero)
+# 	Battlecry: Gain Stealth until your next turn.
+# 	https://hearthstone.gamepedia.com/Valeera_the_Hollow
+# 	"""
+# 	play = None
+# 	pass
+
+
+# class ICC_827e:
+# 	"""
+# 	Shadow Reflection - (Enchantment)
+# 	Always copy your last played card.
+# 	https://hearthstone.gamepedia.com/Shadow_Reflection
+# 	"""
+# 	pass
+
+
+# class ICC_827e3:
+# 	"""
+# 	Veil of Shadows - (Enchantment)
+# 	Stealth until your next turn.
+# 	https://hearthstone.gamepedia.com/Veil_of_Shadows
+# 	"""
+# 	pass
+
+
+# class ICC_827p:
+# 	"""
+# 	Death's Shadow - (HeroPower)
+# 	Passive Hero Power During your turn, add a 'Shadow Reflection' to your hand.
+# 	https://hearthstone.gamepedia.com/Death%27s_Shadow
+# 	"""
+# 	pass
+
+
+# class ICC_827t:
+# 	"""
+# 	Shadow Reflection - (Spell)
+# 	Each time you play a card, transform this into a copy of it.
+# 	https://hearthstone.gamepedia.com/Shadow_Reflection
+# 	"""
+# 	pass
+
+
+# class ICC_850:
+# 	"""
+# 	Shadowblade - (Weapon)
+# 	Battlecry: Your hero is Immune this turn.
+# 	https://hearthstone.gamepedia.com/Shadowblade
+# 	"""
+# 	play = None
+# 	pass
+
+
+# class ICC_850e:
+# 	"""
+# 	Shaded - (Enchantment)
+# 	Immune this turn.
+# 	https://hearthstone.gamepedia.com/Shaded
+# 	"""
+# 	pass
+
+
+# class ICC_910:
+# 	"""
+# 	Spectral Pillager - (Minion)
+# 	Combo: Deal damage equal to the number of other cards you've played this turn.
+# 	https://hearthstone.gamepedia.com/Spectral_Pillager
+# 	"""
+# 	pass
+
+###############################################################################
+##                                                                           ##
+##                                  Priest                                   ##
+##                                                                           ##
+###############################################################################
+
+
+# class ICC_207:
+# 	"""
+# 	Devour Mind - (Spell)
+# 	Copy 3 cards in your opponent's deck and add them to your hand.
+# 	https://hearthstone.gamepedia.com/Devour_Mind
+# 	"""
+# 	pass
+
+
+# class ICC_210:
+# 	"""
+# 	Shadow Ascendant - (Minion)
+# 	At the end of your turn, give another random friendly minion +1/+1.
+# 	https://hearthstone.gamepedia.com/Shadow_Ascendant
+# 	"""
+# 	pass
+
+
+# class ICC_210e:
+# 	"""
+# 	Ascended - (Enchantment)
+# 	Stats increased.
+# 	https://hearthstone.gamepedia.com/Ascended
+# 	"""
+# 	pass
+
+
+# class ICC_213:
+# 	"""
+# 	Eternal Servitude - (Spell)
+# 	Discover a friendly minion that died this game. Summon it.
+# 	https://hearthstone.gamepedia.com/Eternal_Servitude
+# 	"""
+# 	pass
+
+
+# class ICC_214:
+# 	"""
+# 	Obsidian Statue - (Minion)
+# 	Taunt, Lifesteal Deathrattle: Destroy a  random enemy minion.
+# 	https://hearthstone.gamepedia.com/Obsidian_Statue
+# 	"""
+# 	deathrattle = None
+# 	pass
+
+
+# class ICC_215:
+# 	"""
+# 	Archbishop Benedictus - (Minion)
+# 	Battlecry: Shuffle a copy of_your opponent's deck into your deck.
+# 	https://hearthstone.gamepedia.com/Archbishop_Benedictus
+# 	"""
+# 	play = None
+# 	pass
+
+
+# class ICC_235:
+# 	"""
+# 	Shadow Essence - (Spell)
+# 	Summon a 5/5 copy of_a random minion in_your deck.
+# 	https://hearthstone.gamepedia.com/Shadow_Essence
+# 	"""
+# 	pass
+
+
+# class ICC_802:
+# 	"""
+# 	Spirit Lash - (Spell)
+# 	Lifesteal Deal $1 damage to_all_minions.
+# 	https://hearthstone.gamepedia.com/Spirit_Lash
+# 	"""
+# 	pass
+
+
+# class ICC_830:
+# 	"""
+# 	Shadowreaper Anduin - (Hero)
+# 	Battlecry: Destroy all minions with 5 or more_Attack.
+# 	https://hearthstone.gamepedia.com/Shadowreaper_Anduin
+# 	"""
+# 	play = None
+# 	pass
+
+
+# class ICC_830p:
+# 	"""
+# 	Voidform - (HeroPower)
+# 	Hero Power Deal $2 damage. After you play a card, refresh this.
+# 	https://hearthstone.gamepedia.com/Voidform
+# 	"""
+# 	pass
+
+
+# class ICC_849:
+# 	"""
+# 	Embrace Darkness - (Spell)
+# 	Choose an enemy minion. At the start of your turn, gain control of it.
+# 	https://hearthstone.gamepedia.com/Embrace_Darkness
+# 	"""
+# 	pass
+
+###############################################################################
+##                                                                           ##
+##                                  Paladin                                  ##
+##                                                                           ##
+###############################################################################
+
+
+# class ICC_034:
+# 	"""
+# 	Arrogant Crusader - (Minion)
+# 	Deathrattle: If it's your opponent's turn, summon a 2/2 Ghoul.
+# 	https://hearthstone.gamepedia.com/Arrogant_Crusader
+# 	"""
+# 	deathrattle = None
+# 	pass
+
+
+# class ICC_039:
+# 	"""
+# 	Dark Conviction - (Spell)
+# 	Set a minion's Attack and Health to 3.
+# 	https://hearthstone.gamepedia.com/Dark_Conviction
+# 	"""
+# 	pass
+
+
+# class ICC_071:
+# 	"""
+# 	Light's Sorrow - (Weapon)
+# 	After a friendly minion loses Divine Shield, gain +1 Attack.
+# 	https://hearthstone.gamepedia.com/Light%27s_Sorrow
+# 	"""
+# 	pass
+
+
+# class ICC_071e:
+# 	"""
+# 	Dying Light - (Enchantment)
+# 	Increased Attack.
+# 	https://hearthstone.gamepedia.com/Dying_Light
+# 	"""
+# 	pass
+
+
+# class ICC_244:
+# 	"""
+# 	Desperate Stand - (Spell)
+# 	Give a minion "Deathrattle: Return this to life with 1 Health."
+# 	https://hearthstone.gamepedia.com/Desperate_Stand
+# 	"""
+# 	pass
+
+
+# class ICC_244e:
+# 	"""
+# 	Redeemed - (Enchantment)
+# 	Deathrattle: Return this to life with 1 Health.
+# 	https://hearthstone.gamepedia.com/Redeemed
+# 	"""
+# 	pass
+
+
+# class ICC_245:
+# 	"""
+# 	Blackguard - (Minion)
+# 	Whenever your hero is healed, deal that much damage to a random enemy minion.
+# 	https://hearthstone.gamepedia.com/Blackguard
+# 	"""
+# 	pass
+
+
+# class ICC_801:
+# 	"""
+# 	Howling Commander - (Minion)
+# 	Battlecry: Draw a Divine_Shield minion from your deck.
+# 	https://hearthstone.gamepedia.com/Howling_Commander
+# 	"""
+# 	play = None
+# 	pass
+
+
+# class ICC_829:
+# 	"""
+# 	Uther of the Ebon Blade - (Hero)
+# 	Battlecry: Equip a 5/3 Lifesteal weapon.
+# 	https://hearthstone.gamepedia.com/Uther_of_the_Ebon_Blade
+# 	"""
+# 	play = None
+# 	pass
+
+
+# class ICC_858:
+# 	"""
+# 	Bolvar, Fireblood - (Minion)
+# 	Divine Shield After a friendly minion loses_Divine Shield, gain +2 Attack.
+# 	https://hearthstone.gamepedia.com/Bolvar%2C_Fireblood
+# 	"""
+# 	pass
+
+
+# class ICC_858e:
+# 	"""
+# 	Fading Light - (Enchantment)
+# 	Increased Attack.
+# 	https://hearthstone.gamepedia.com/Fading_Light
+# 	"""
+# 	pass
+
+###############################################################################
+##                                                                           ##
+##                                   Mage                                    ##
+##                                                                           ##
+###############################################################################
+
+
+# class ICCA04_002:
+# 	"""
+# 	Unchained Magic - (Spell)
+# 	Deal $3 damage to the enemy hero for each spell in their hand.
+# 	https://hearthstone.gamepedia.com/Unchained_Magic
+# 	"""
+# 	pass
+
+
+# class ICC_068:
+# 	"""
+# 	Ice Walker - (Minion)
+# 	Your Hero Power also Freezes the target.
+# 	https://hearthstone.gamepedia.com/Ice_Walker
+# 	"""
+# 	pass
+
+
+# class ICC_069:
+# 	"""
+# 	Ghastly Conjurer - (Minion)
+# 	Battlecry: Add a 'Mirror Image' spell to your hand.
+# 	https://hearthstone.gamepedia.com/Ghastly_Conjurer
+# 	"""
+# 	play = None
+# 	pass
+
+
+class ICC_082:
+	"""
+	Frozen Clone - (Spell)
+	Secret: After your opponent plays a minion, add two copies of it to_your hand.
+	https://hearthstone.gamepedia.com/Frozen_Clone
+	"""
+	secret = Play(OPPONENT, MINION).after(
+		Reveal(SELF), Give(CONTROLLER, Copy(Play.CARD)) * 2
+	)
+
+
+# class ICC_083:
+# 	"""
+# 	Doomed Apprentice - (Minion)
+# 	Your opponent's spells cost (1) more.
+# 	https://hearthstone.gamepedia.com/Doomed_Apprentice
+# 	"""
+# 	pass
+
+
+# class ICC_086:
+# 	"""
+# 	Glacial Mysteries - (Spell)
+# 	Put one of each Secret from your deck into the battlefield.
+# 	https://hearthstone.gamepedia.com/Glacial_Mysteries
+# 	"""
+# 	pass
+
+
+# class ICC_252:
+# 	"""
+# 	Coldwraith - (Minion)
+# 	Battlecry: If an enemy is Frozen, draw a card.
+# 	https://hearthstone.gamepedia.com/Coldwraith
+# 	"""
+# 	play = None
+# 	pass
+
+
+# class ICC_823:
+# 	"""
+# 	Simulacrum - (Spell)
+# 	Copy the lowest Cost minion in your hand.
+# 	https://hearthstone.gamepedia.com/Simulacrum
+# 	"""
+# 	pass
+
+
+# class ICC_833:
+# 	"""
+# 	Frost Lich Jaina - (Hero)
+# 	Battlecry: Summon a 3/6 Water Elemental. Your Elementals have Lifesteal this game.
+# 	https://hearthstone.gamepedia.com/Frost_Lich_Jaina
+# 	"""
+# 	play = None
+# 	pass
+
+
+# class ICC_833t:
+# 	"""
+# 	Water Elemental - (Minion)
+# 	Freeze any character damaged by this minion.
+# 	https://hearthstone.gamepedia.com/Water_Elemental
+# 	"""
+# 	pass
+
+
+# class ICC_836:
+# 	"""
+# 	Breath of Sindragosa - (Spell)
+# 	Deal $2 damage to a random enemy minion and Freeze it.
+# 	https://hearthstone.gamepedia.com/Breath_of_Sindragosa
+# 	"""
+# 	pass
+
+
+# class ICC_838:
+# 	"""
+# 	Sindragosa - (Minion)
+# 	Battlecry: Summon two 0/1 Frozen Champions.
+# 	https://hearthstone.gamepedia.com/Sindragosa
+# 	"""
+# 	play = None
+# 	pass
+
+
+# class ICC_838t:
+# 	"""
+# 	Frozen Champion - (Minion)
+# 	Deathrattle: Add a random Legendary minion to your hand.
+# 	https://hearthstone.gamepedia.com/Frozen_Champion
+# 	"""
+# 	deathrattle = None
+# 	pass
+
+###############################################################################
+##                                                                           ##
+##                                  Hunter                                   ##
+##                                                                           ##
+###############################################################################
+
+
+# class ICC_021:
+# 	"""
+# 	Exploding Bloatbat - (Minion)
+# 	Deathrattle: Deal 2 damage to all enemy minions.
+# 	https://hearthstone.gamepedia.com/Exploding_Bloatbat
+# 	"""
+# 	deathrattle = None
+# 	pass
+
+
+# class ICC_049:
+# 	"""
+# 	Toxic Arrow - (Spell)
+# 	Deal $2 damage to a minion. If it survives, give it Poisonous.
+# 	https://hearthstone.gamepedia.com/Toxic_Arrow
+# 	"""
+# 	pass
+
+
+# class ICC_052:
+# 	"""
+# 	Play Dead - (Spell)
+# 	Trigger a friendly minion's Deathrattle.
+# 	https://hearthstone.gamepedia.com/Play_Dead
+# 	"""
+# 	pass
+
+
+# class ICC_200:
+# 	"""
+# 	Venomstrike Trap - (Spell)
+# 	Secret: When one of your minions is attacked, summon a 2/3_Poisonous Cobra.
+# 	https://hearthstone.gamepedia.com/Venomstrike_Trap
+# 	"""
+# 	pass
+
+
+# class ICC_204:
+# 	"""
+# 	Professor Putricide - (Minion)
+# 	After you play a Secret, put a random Hunter Secret into the battlefield.
+# 	https://hearthstone.gamepedia.com/Professor_Putricide
+# 	"""
+# 	pass
+
+
+# class ICC_243:
+# 	"""
+# 	Corpse Widow - (Minion)
+# 	Your Deathrattle cards cost_(2) less.
+# 	https://hearthstone.gamepedia.com/Corpse_Widow
+# 	"""
+# 	pass
+
+
+# class ICC_415:
+# 	"""
+# 	Stitched Tracker - (Minion)
+# 	Battlecry: Discover a copy of a minion in your deck.
+# 	https://hearthstone.gamepedia.com/Stitched_Tracker
+# 	"""
+# 	play = None
+# 	pass
+
+
+# class ICC_825:
+# 	"""
+# 	Abominable Bowman - (Minion)
+# 	Deathrattle: Summon a random friendly Beast that died this game.
+# 	https://hearthstone.gamepedia.com/Abominable_Bowman
+# 	"""
+# 	deathrattle = None
+# 	pass
+
+
+# class ICC_828:
+# 	"""
+# 	Deathstalker Rexxar - (Hero)
+# 	Battlecry: Deal 2 damage to all enemy minions.
+# 	https://hearthstone.gamepedia.com/Deathstalker_Rexxar
+# 	"""
+# 	play = None
+# 	pass
+
+
+# class ICC_828e:
+# 	"""
+# 	Stitched - (Enchantment)
+# 	{0} and {1}.
+# 	https://hearthstone.gamepedia.com/Stitched
+# 	"""
+# 	pass
+
+
+# class ICC_828t:
+# 	"""
+# 	Zombeast - (Minion)
+# 	{0} {1}
+# 	https://hearthstone.gamepedia.com/Zombeast
+# 	"""
+# 	pass
+
+###############################################################################
+##                                                                           ##
+##                                   Druid                                   ##
+##                                                                           ##
+###############################################################################
+
+
+# class ICC_047:
+# 	"""
+# 	Fatespinner - (Minion)
+# 	Choose a Deathrattle (Secretly) - Deal 3 damage to all minions; or Give them +2/+2.
+# 	https://hearthstone.gamepedia.com/Fatespinner
+# 	"""
+# 	pass
+
+
+# class ICC_047a:
+# 	"""
+# 	Growth - (Spell)
+# 	Deathrattle: Give all minions +2/+2.
+# 	https://hearthstone.gamepedia.com/Growth
+# 	"""
+# 	pass
+
+
+# class ICC_047b:
+# 	"""
+# 	Decay - (Spell)
+# 	Deathrattle: Deal 3 damage to all minions.
+# 	https://hearthstone.gamepedia.com/Decay
+# 	"""
+# 	pass
+
+
+# class ICC_047e:
+# 	"""
+# 	Growth - (Enchantment)
+# 	+2/+2.
+# 	https://hearthstone.gamepedia.com/Growth
+# 	"""
+# 	pass
+
+
+# class ICC_047t:
+# 	"""
+# 	Fatespinner - (Minion)
+# 	Secret Deathrattle: Deal 3 damage to all minions; or Give them +2/+2.@Secret Deathrattle: Give +2/+2 to all minions.@Secret Deathrattle: Deal 3 damage to all minions.
+# 	https://hearthstone.gamepedia.com/Fatespinner
+# 	"""
+# 	deathrattle = None
+# 	pass
+
+
+# class ICC_047t2:
+# 	"""
+# 	Fatespinner - (Minion)
+# 	Deathrattle: Deal 3 damage to all minions and give them +2/+2.
+# 	https://hearthstone.gamepedia.com/Fatespinner
+# 	"""
+# 	deathrattle = None
+# 	pass
+
+
+# class ICC_050:
+# 	"""
+# 	Webweave - (Spell)
+# 	Summon two 1/2 Poisonous Spiders.
+# 	https://hearthstone.gamepedia.com/Webweave
+# 	"""
+# 	pass
+
+
+# class ICC_051:
+# 	"""
+# 	Druid of the Swarm - (Minion)
+# 	Choose One - Transform into a 1/2 with Poisonous; or a 1/5 with Taunt.
+# 	https://hearthstone.gamepedia.com/Druid_of_the_Swarm
+# 	"""
+# 	pass
+
+
+# class ICC_051b:
+# 	"""
+# 	Scarab Form - (Spell)
+# 	+3 Health and Taunt.
+# 	https://hearthstone.gamepedia.com/Scarab_Form
+# 	"""
+# 	pass
+
+
+# class ICC_054:
+# 	"""
+# 	Spreading Plague - (Spell)
+# 	Summon a 1/5 Scarab with Taunt. If your_opponent has more minions, cast this again.
+# 	https://hearthstone.gamepedia.com/Spreading_Plague
+# 	"""
+# 	pass
+
+
+# class ICC_079:
+# 	"""
+# 	Gnash - (Spell)
+# 	Give your hero +3_Attack this turn. Gain 3 Armor.
+# 	https://hearthstone.gamepedia.com/Gnash
+# 	"""
+# 	pass
+
+
+# class ICC_079e:
+# 	"""
+# 	Gnash - (Enchantment)
+# 	+3 Attack this turn.
+# 	https://hearthstone.gamepedia.com/Gnash
+# 	"""
+# 	pass
+
+
+# class ICC_085:
+# 	"""
+# 	Ultimate Infestation - (Spell)
+# 	Deal $5 damage. Draw 5 cards. Gain 5 Armor. Summon a 5/5 Ghoul.
+# 	https://hearthstone.gamepedia.com/Ultimate_Infestation
+# 	"""
+# 	pass
+
+
+# class ICC_807:
+# 	"""
+# 	Strongshell Scavenger - (Minion)
+# 	Battlecry: Give your Taunt minions +2/+2.
+# 	https://hearthstone.gamepedia.com/Strongshell_Scavenger
+# 	"""
+# 	play = None
+# 	pass
+
+
+# class ICC_808:
+# 	"""
+# 	Crypt Lord - (Minion)
+# 	Taunt After you summon a minion,  gain +1 Health.
+# 	https://hearthstone.gamepedia.com/Crypt_Lord
+# 	"""
+# 	pass
+
+
+# class ICC_808e:
+# 	"""
+# 	Might of Ner'ub - (Enchantment)
+# 	Increased Health.
+# 	https://hearthstone.gamepedia.com/Might_of_Ner%27ub
+# 	"""
+# 	pass
+
+
+# class ICC_832:
+# 	"""
+# 	Malfurion the Pestilent - (Hero)
+# 	Choose One - Summon 2 Poisonous Spiders; or 2 Scarabs with Taunt.
+# 	https://hearthstone.gamepedia.com/Malfurion_the_Pestilent
+# 	"""
+# 	pass
+
+
+# class ICC_832a:
+# 	"""
+# 	Scarab Plague - (Spell)
+# 	Summon two 1/5 Scarabs with Taunt.
+# 	https://hearthstone.gamepedia.com/Scarab_Plague
+# 	"""
+# 	pass
+
+
+# class ICC_832b:
+# 	"""
+# 	Spider Plague - (Spell)
+# 	Summon two 1/2 Spiders with Poisonous.
+# 	https://hearthstone.gamepedia.com/Spider_Plague
+# 	"""
+# 	pass
+
+
+# class ICC_832pa:
+# 	"""
+# 	Scarab Shell - (Spell)
+# 	+3 Armor.
+# 	https://hearthstone.gamepedia.com/Scarab_Shell
+# 	"""
+# 	pass
+
+
+# class ICC_832pb:
+# 	"""
+# 	Spider Fangs - (Spell)
+# 	+3 Attack.
+# 	https://hearthstone.gamepedia.com/Spider_Fangs
+# 	"""
+# 	pass
+
+
+# class ICC_835:
+# 	"""
+# 	Hadronox - (Minion)
+# 	Deathrattle: Summon your Taunt minions that died this game.
+# 	https://hearthstone.gamepedia.com/Hadronox
+# 	"""
+# 	deathrattle = None
+# 	pass
+
+###############################################################################
+##                                                                           ##
+##                               Death Knight                                ##
+##                                                                           ##
+###############################################################################
+
+
+# class ICCA01_003:
+# 	"""
+# 	Blood Tap - (Spell)
+# 	The next spell you cast this turn costs (0).
+# 	https://hearthstone.gamepedia.com/Blood_Tap
+# 	"""
+# 	pass
+
+
+# class ICCA01_004:
+# 	"""
+# 	Army of the Dead - (Spell)
+# 	Summon seven 3/3 Ghouls with Taunt.
+# 	https://hearthstone.gamepedia.com/Army_of_the_Dead
+# 	"""
+# 	pass
+
+
+# class ICCA08_020:
+# 	"""
+# 	Frostmourne - (Weapon)
+# 	Battlecry: Remove all your minions. Deathrattle: Resummon them.
+# 	https://hearthstone.gamepedia.com/Frostmourne
+# 	"""
+# 	play = None
+# 	pass
+
+
+# class ICCA08_021:
+# 	"""
+# 	The True Lich - (Spell)
+# 	Set the enemy hero's remaining Health to 1.
+# 	https://hearthstone.gamepedia.com/The_True_Lich
+# 	"""
+# 	pass
+
+
+# class ICCA08_022:
+# 	"""
+# 	Fallen Champions - (Spell)
+# 	Whenever an enemy dies this game, take control of it.
+# 	https://hearthstone.gamepedia.com/Fallen_Champions
+# 	"""
+# 	pass
+
+
+# class ICCA08_023:
+# 	"""
+# 	Necrotic Plague - (Spell)
+# 	Set minions in your opponent's hand and deck to 1/1.
+# 	https://hearthstone.gamepedia.com/Necrotic_Plague
+# 	"""
+# 	pass
+
+
+# class ICCA08_023e:
+# 	"""
+# 	Plagued - (Enchantment)
+# 	1/1.
+# 	https://hearthstone.gamepedia.com/Plagued
+# 	"""
+# 	pass
+
+
+# class ICCA08_024:
+# 	"""
+# 	The Hunted - (Spell)
+# 	Deal $2 damage to your opponent for each minion in their deck.
+# 	https://hearthstone.gamepedia.com/The_Hunted
+# 	"""
+# 	pass
+
+
+# class ICCA08_025:
+# 	"""
+# 	Purge the Weak - (Spell)
+# 	Destroy all enemy minions that cost (3) or less, wherever they are.
+# 	https://hearthstone.gamepedia.com/Purge_the_Weak
+# 	"""
+# 	pass
+
+
+# class ICCA08_026:
+# 	"""
+# 	Soul Reaper - (Spell)
+# 	Deal $2 damage for each duplicate in opponent’s deck.
+# 	https://hearthstone.gamepedia.com/Soul_Reaper
+# 	"""
+# 	pass
+
+
+# class ICCA08_027:
+# 	"""
+# 	The True King - (Spell)
+# 	Gain 100 Armor.
+# 	https://hearthstone.gamepedia.com/The_True_King
+# 	"""
+# 	pass
+
+
+# class ICCA08_028:
+# 	"""
+# 	The Price of Power - (Spell)
+# 	Destroy every spell in your opponent's hand and deck.
+# 	https://hearthstone.gamepedia.com/The_Price_of_Power
+# 	"""
+# 	pass
+
+
+# class ICCA08_029:
+# 	"""
+# 	Shut up, Priest - (Spell)
+# 	Your opponent can't emote.
+# 	https://hearthstone.gamepedia.com/Shut_up%2C_Priest
+# 	"""
+# 	pass
+
+
+# class ICCA08_030p:
+# 	"""
+# 	Remorseless Winter - (HeroPower)
+# 	Hero Power Deal @ damage to the enemy hero. +1 Damage each time.
+# 	https://hearthstone.gamepedia.com/Remorseless_Winter
+# 	"""
+# 	pass
+
+
+# class ICC_314t1:
+# 	"""
+# 	Frostmourne - (Weapon)
+# 	Deathrattle: Summon every minion killed by this weapon.
+# 	https://hearthstone.gamepedia.com/Frostmourne
+# 	"""
+# 	deathrattle = None
+# 	pass
+
+
+# class ICC_314t2:
+# 	"""
+# 	Army of the Dead - (Spell)
+# 	Remove the top 5 cards of your deck. Summon any minions removed.
+# 	https://hearthstone.gamepedia.com/Army_of_the_Dead
+# 	"""
+# 	pass
+
+
+# class ICC_314t3:
+# 	"""
+# 	Doom Pact - (Spell)
+# 	Destroy all minions.  Remove the top card  from your deck for each minion destroyed.
+# 	https://hearthstone.gamepedia.com/Doom_Pact
+# 	"""
+# 	pass
+
+
+# class ICC_314t4:
+# 	"""
+# 	Death Grip - (Spell)
+# 	Steal a minion from your opponent's deck and add it to your hand.
+# 	https://hearthstone.gamepedia.com/Death_Grip
+# 	"""
+# 	pass
+
+
+# class ICC_314t5:
+# 	"""
+# 	Death Coil - (Spell)
+# 	Deal $5 damage to an enemy, or restore #5 Health to a friendly character.
+# 	https://hearthstone.gamepedia.com/Death_Coil
+# 	"""
+# 	pass
+
+
+# class ICC_314t6:
+# 	"""
+# 	Obliterate - (Spell)
+# 	Destroy a minion. Your hero takes damage equal to its Health.
+# 	https://hearthstone.gamepedia.com/Obliterate
+# 	"""
+# 	pass
+
+
+# class ICC_314t7:
+# 	"""
+# 	Anti-Magic Shell - (Spell)
+# 	Give your minions +2/+2 and "Can't be targeted by spells or Hero Powers."
+# 	https://hearthstone.gamepedia.com/Anti-Magic_Shell
+# 	"""
+# 	pass
+
+
+# class ICC_314t8:
+# 	"""
+# 	Death and Decay - (Spell)
+# 	Deal $3 damage to all enemies.
+# 	https://hearthstone.gamepedia.com/Death_and_Decay
+# 	"""
+# 	pass
